# faster `hypot(::IEEEFloat...)`