`hypot(::IEEEFloat...)` can be made faster for 3+ arguments.

```julia
using BenchmarkTools

t3 = ntuple(_->randn(), 3)
t4 = ntuple(_->randn(), 4)
t8 = ntuple(_->randn(), 8)
t16 = ntuple(_->randn(), 16)

@btime hypot($t3...)
@btime hypot($t4...)
@btime hypot($t8...)
@btime hypot($t16...)

# master
  9.810 ns (0 allocations: 0 bytes)
  10.110 ns (0 allocations: 0 bytes)
  19.759 ns (0 allocations: 0 bytes)
  50.203 ns (0 allocations: 0 bytes)

# SHA 1a0340c
  7.808 ns (0 allocations: 0 bytes)
  8.008 ns (0 allocations: 0 bytes)
  10.010 ns (0 allocations: 0 bytes)
  14.529 ns (0 allocations: 0 bytes)
```

The notable savings here come from casting to integers before computing `maximum(abs,...)` (same result but faster comparisons) and re-associating the final sum (see #48129).