@mikmoore 2023-01-05
> Does this help with #47917?

It appears that issue should be closed on the current `master` -- probably by #44357. You can see from the timings presented above that 3 not slower than 4 (although it's only marginally faster) on both `master` and this PR.