@giordano 2023-01-06
> It appears that issue should be closed on the current master -- probably by #44357.

#44357 is already in v1.9, where #47917 was reported on.