@KristofferC 2023-01-06
I don't see the issue having been reprod on beta2.