@mikmoore 2023-01-07
In any case, the issue was fixed before any efforts from this PR (or otherwise never existed on my architecture). If someone's ambitious, they can try to find the responsible PR and see if it'll be accepted as a backport to 1.9.