@IanButterworth 2023-01-28
I think this broke build. Was CI green on it?
```
math.jl
error during bootstrap:
LoadError("sysimg.jl", 3, LoadError("Base.jl", 328, LoadError("math.jl", 844, LoadError("math.jl", 851, UndefVarError(Symbol("@fastmath"))))))
ijl_undefined_var_error at /home/ian/Documents/GitHub/julia/src/rtutils.c:132
```