@DilumAluthge 2023-01-28
> Was CI green on it?

No.

See: https://buildkite.com/julialang/julia-master/builds?branch=mikmoore%3Ahypot