@mikmoore 2023-02-07
So I'm guessing the issue here is that `@fastmath` is not defined at the point this code is loaded. It's significant to the performance here, due to #48129. Does someone have a suggestion for how to resolve this?