@giordano 2023-02-07
Move hypot to another file which is included in Base after fastmath?